refactor: 기록 살리기 리마인더 판단 로직을 RecoveryReminderPolicy로 분리
